fix : update to cut floating point

+++ src/main/java/wecandoeverything/ledgerly/service/ReceiptScanService.java
@@ -92,7 +92,7 @@ public class ReceiptScanService {
         return result.toBuilder()
                 .possibleDuplicate(true)
                 .duplicateNote("비슷한 요청" + result.getMerchant() +
-                        " (₩`" + result.getAmount() + ") 이 이미 있습니다.")
+                        " (₩`" + result.getAmount().intValue() + ") 이 이미 있습니다.")
                 .build();
     }
